refactor(day-diet): split day diet use cases into granular files
