+++ src/modules/diet/day-diet/application/usecases/dayState.ts
@@ -1,0 +1,7 @@
+import { dayDietStore } from '~/modules/diet/day-diet/application/dayDietStore'
+
+export const targetDay = dayDietStore.targetDay
+export const setTargetDay = dayDietStore.setTargetDay
+
+export const currentToday = dayDietStore.currentToday
+export const currentDayDiet = dayDietStore.currentDayDiet
