+++ src/sections/common/components/TargetDayPicker.tsx
@@ -3,7 +3,7 @@ import { Suspense } from 'solid-js'
 import {
   setTargetDay,
   targetDay,
-} from '~/modules/diet/day-diet/application/dayDietUseCases'
+} from '~/modules/diet/day-diet/application/usecases/dayState'
 import { type DateValueType } from '~/sections/datepicker/types'
 import { lazyImport } from '~/shared/solid/lazyImport'
 import { getTodayYYYYMMDD, stringToDate } from '~/shared/utils/date/dateUtils'
