+++ src/sections/day-diet/components/DayMacros.tsx
@@ -1,6 +1,6 @@
 import { createMemo, Show } from 'solid-js'
 
-import { currentDayDiet } from '~/modules/diet/day-diet/application/dayDietUseCases'
+import { currentDayDiet } from '~/modules/diet/day-diet/application/usecases/dayState'
 import { type DayDiet } from '~/modules/diet/day-diet/domain/dayDiet'
 import {
   createMacroNutrients,
